requirements.txt
keras==2.11.0
tensorflow==2.11.0
matplotlib==3.6.3
numpy==1.24.2
Pillow==9.4.0
scikit_learn==1.2.1
